+++ mesos_slave/tests/test_check.py
@@ -1,49 +1,53 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-# (C) Datadog, Inc. 2018
-# All rights reserved
-# Licensed under a 3-clause BSD style license (see LICENSE)
-
-import json
-import os
-
 import pytest
 from six import iteritems
 
-from datadog_checks.mesos_slave import MesosSlave
 
-CHECK_NAME = 'mesos_master'
-FIXTURE_DIR = os.path.join(os.path.dirname(__file__), 'fixtures')
+# TODO: Fix docker-compose.yml networking set up
+#
+# @pytest.mark.usefixtures('dd_environment')
+# def test_check(check, instance, aggregator):
+#     check.check(instance)
+#     metrics = {}
+#     for d in (
+#         check.SLAVE_TASKS_METRICS,
+#         check.SYSTEM_METRICS,
+#         check.SLAVE_RESOURCE_METRICS,
+#         check.SLAVE_EXECUTORS_METRICS,
+#         check.STATS_METRICS,
+#     ):
+#         metrics.update(d)
+
+#     for _, v in iteritems(check.TASK_METRICS):
+#         aggregator.assert_metric(v[0])
+#     for _, v in iteritems(metrics):
+#         aggregator.assert_metric(v[0])
+
+#     service_check_tags = [
+#         'instance:mytag1',
+#         'mesos_cluster:test',
+#         'mesos_node:slave',
+#         'mesos_pid:slave(1)@127.0.0.1:5051',
+#         'task_name:hello',
+#     ]
+#     aggregator.assert_service_check('hello.ok', tags=service_check_tags, count=1, status=check.OK)
 
 
-def read_fixture(name):
-    with open(os.path.join(FIXTURE_DIR, name)) as f:
-        return f.read()
-
-
-@pytest.fixture
-def check():
-    check = MesosSlave(CHECK_NAME, {}, {})
-    check._get_stats = lambda v, x, y, z: json.loads(read_fixture('stats.json'))
-    check._get_state = lambda v, x, y, z: json.loads(read_fixture('state.json'))
-
-    return check
-
-
-def test_check(check, instance, aggregator):
-    check.check(instance)
+def test_fixtures(check_mock, instance, aggregator):
+    check_mock.check(instance)
     metrics = {}
     for d in (
-        check.SLAVE_TASKS_METRICS,
-        check.SYSTEM_METRICS,
-        check.SLAVE_RESOURCE_METRICS,
-        check.SLAVE_EXECUTORS_METRICS,
-        check.STATS_METRICS,
+        check_mock.SLAVE_TASKS_METRICS,
+        check_mock.SYSTEM_METRICS,
+        check_mock.SLAVE_RESOURCE_METRICS,
+        check_mock.SLAVE_EXECUTORS_METRICS,
+        check_mock.STATS_METRICS,
     ):
         metrics.update(d)
 
-    for _, v in iteritems(check.TASK_METRICS):
+    for _, v in iteritems(check_mock.TASK_METRICS):
         aggregator.assert_metric(v[0])
     for _, v in iteritems(metrics):
         aggregator.assert_metric(v[0])
@@ -55,4 +59,4 @@ def test_check(check, instance, aggregator):
         'mesos_pid:slave(1)@127.0.0.1:5051',
         'task_name:hello',
     ]
-    aggregator.assert_service_check('hello.ok', tags=service_check_tags, count=1, status=MesosSlave.OK)
+    aggregator.assert_service_check('hello.ok', tags=service_check_tags, count=1, status=check_mock.OK)
