+++ mesos_slave/tox.ini
@@ -2,12 +2,15 @@
 minversion = 2.0
 basepython = py37
 envlist =
-    py{27,37}-unit
+    py{27,37}
 
 [testenv]
 dd_check_style = true
 usedevelop = true
 platform = linux|darwin|win32
+passenv =
+    DOCKER*
+    COMPOSE*
 deps =
     -e../datadog_checks_base[deps]
     -rrequirements-dev.txt
